Crate etcd

Crate etcd provides a client for the etcd API.

All of the public types are rexported and available directly from the crate root. Client is the entry point for all API calls.

Structs

Client
API client for etcd.

CountStats
Statistics about the number of successful and failed Raft RPC requests to an etcd node.

FollowerStats
Statistics on the health of a single etcd follower node.

KeySpaceInfo
Information about the result of a successful key space operation.

LatencyStats
Statistics about the network latency to an etcd node.

LeaderStats
Statistics about an etcd cluster leader.

Node
An etcd key-value pair or directory.

VersionInfo
Versions of the etcd cluster and server.

Enums

Error
An error returned by Client when an API call fails.

Type Definitions

KeySpaceResult
Returned by key space API calls.