Crate etcd (0.4.0)

Crate etcd provides a client for etcd, a distributed key-value store from CoreOS.

Client is the entry point for all API calls. Types for primary key space operations are reexported to the crate root. Other categories of operations have public types in their respective modules.

Reexports

pub use client::Client;

pub use error::{EtcdResult, Error};

pub use keys::{KeySpaceInfo, KeySpaceResult};

Modules

client
Contains the etcd client. All API calls are made via the client.

error
Contains etcd error types.

keys
Types for the primary key space operations.

stats
Types for statistics operations.

version
Types for the version endpoint.